Split the Logicbuilder files a bit. Also made the MApping table collapsible
