+++ frontend/js/pages/admin/LogicBuilder/DOMGuide.txt
@@ -1,0 +1,5 @@
+ROOT
+└── LogicBuilder
+    └── CQuestionnaireSelection
+        ├── CLogicWorkshop
+        └── CQuestionnaireTable
